Allow applications to customize how DARTSESSIONID is generated
In a scenario that we serves a domain with multiple Dart servers, we need to share the same value of DARTSESSIONID among these servers. With the current implementation, a new value is generated each time a user visits a page served by a different server. It ends up a lot of sessions stores in each server if user jumps around these pages. Not mention the effort to fail over these sessions.

One straightforward solution is to allow a customized session ID generator. That is, Dart SDK allows developers to plug in a generator to generate ID the way application needs it. Of course, it requires some additional API for plugin.

Another solution is like this pull request: generate the DARTSESSIONID cookie earlier. Thus, application can play around the cookie to have the effect it wants -- that is, browser can use the same session ID while different servers can has its own session ID. That is, application can implement a mapping to use same DARTSESSIONID. IMO, it doesn't change the logic nor introduce new API.