Use known FileState from FileSystemState directly.
We check all known files now, so we don't need looking up FileState(s)
by paths. It is expensive when we do it for thousand files in multiple
drivers.

Stopwatch is also not free, so now we check it every 100 files.

This makes _FilesReferencingNameTask about 4 times faster.
In total search of an element in a priority file is 2 times faster.

R=brianwilkerson@google.com

Change-Id: I2738df8035f84d20a8b7e1b6d3040d4461c2609d
Reviewed-on: https://dart-review.googlesource.com/58141
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>