Treat objects with .call as "subtypes" of function types
There were a pair of TODOs in constraint gathering in type inference,
where objects with .call methods should be considered as if they were
subtypes of function types.  Fix them.

This is a partial fix for
https://github.com/dart-lang/sdk/issues/33298 because it does not
yet handle the case of generic .call methods.

Bug: https://github.com/dart-lang/sdk/issues/33298
Change-Id: Ib9803db9a152cdc6f2108e717bb0934c38016a94
Reviewed-on: https://dart-review.googlesource.com/58206
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>