[VM] Clear out return-value-register from call-to-runtime-stub, it needs to be GC-safe for lazy deopts
When the slow path code of StackOverflowInstr is calling out to runtime
and the runtime causes a lazy-deopt, the lazy deopt stub will treat the value in the
return-value-register as GC-safe and the value will be visited during GCs.

This CL changes the call-to-runtime-stub to always put 0 into the
return-value-register.

This should fix the flaky vm-linux-debug-x64-reload-be builder.

Change-Id: Ibea3630b8b9fa5de2b96134725df0017c09282bd
Reviewed-on: https://dart-review.googlesource.com/60661
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>