Modify Factory API to allow forward references to incomplete local functions and variables.
To allow for local functions that call themselves recursively, as well
as variable declarations that are (erroneously) referred to prior to
their first use, the Factory API needs to be adjusted so that it's
possible for a local function (or local variable) to be referred to
prior to its body (or initializer) being built.

The analyzer and the kernel have different ways of referring to local
functions and variables; in the kernel the reference
(e.g. VariableGet) points directly to the declaration
(VariableDeclaration).  In the analyzer both the reference and the
declaration point to a FunctionElement (or LocalVariableElement).  So
we want the Factory API to support both approaches.

This CL introduces the idea of a "lemma" (borrowed from linguistics).
In linguistics, a lemma is the mental binding site to which words and
meanings are attached.  In the Factory API, a lemma is an object that
can be created before the function's body (or the variable's
initializer) is available.  Thereafter, the declaration and the
references can be bound to the lemma in any order.

In the analyzer implementation, the lemma will be represented by a
FunctionElement or LocalVariableElement.  (Note however that for the
moment we use a temporary object because the ResolutionStorer is not
yet creating elements and AST nodes).  In the kernel implementation,
the VariableDeclaration node itself is used as the lemma, so the act
of "binding" the lemma to a declaration will do nothing more than
storing the body or initializer, and returning the same object.

Note that there are other syntactic constructs that can declare local
variables (closure parameters, for loops, and catch clauses).  I'll
hook them up to the lemma mechanism in a future CL.

Also a future CL will introduce the same mechanism for labeled
statements (which have a similar need for references to be built
before the labeled statement is completed).

Change-Id: Ifa3b04dda7ce25d115ad58a471d164b4c24d539c
Reviewed-on: https://dart-review.googlesource.com/61720
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>