Remove code awaiting void results to unblock #33145
It's possible that the code relies upon this async gap. However, I
imagine it isn't, and that the await was merely added in these places
because it looks like it would be an asynchronous operation. In any
case, it was not waiting on anything before, and so any reliance upon
any async gaps here would be strange.

Change-Id: I49c3be4dcaaa18361ef60c7f9e31999db3fc090b
Reviewed-on: https://dart-review.googlesource.com/60704
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>