Replace strstr with strncmp looking for path separator
Using strstr in a loop is O(n^2) with n being the length of the path,
whereas using strncmp is O(n*m) with n being the length of the path
and m is the length of the separator (which is always 1).
Effectively this turns the quadratic behavior linear.

Change-Id: If60d02c33eeca92ca2ff52e0df97d0bb80fe57e0
Reviewed-on: https://dart-review.googlesource.com/62144
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>