Detect when a class has a member that conflicts with a static one
In Dart 2 classes are not allowed to have an instance method, getter,
or setter that conflicts with a static method, getter, or setter
declared in the class.  Detect this at the start of top-level type
inference when we build an explicit interface for every class.

Change-Id: I39c1887a07fde0c40367d8f243a37f06dfbba353
Reviewed-on: https://dart-review.googlesource.com/60581
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Aske Simon Christensen <askesc@google.com>