Fix incorrect handling of non-generic typedefs.
Currently a typedef is represented by two elements: an "outer" element
that points to the typedef and has type GenericTypeAliasElementImpl,
and an "inner" element that points to the RHS of the typedef and has
type GenericFunctionTypeElementImpl.  GenericTypeAliasElementImpl.type
is never used and handles generic parameters incorrectly;
GenericFunctionTypeElementImpl.type should always be used instead.

I plan to clean all this up as part of #33641, but this CL is
sufficient to get tests passing.

Change-Id: Ib2d9af9294f4f9d94bfbbeff9aaae711f8cdad1d
Reviewed-on: https://dart-review.googlesource.com/62540
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>