In resynthesis tests, do not attempt to distinguish missing bound from Object.
According to the spec, a missing type parameter bound is
indistinguishable from a type parameter bound of `Object`.  Kernel and
the front end are consistent with the specification; the analyzer is
not.  (In the analyzer, TypeParameterElement.bound returns `null` if
no bound was specified, but a DartType of `Object` if `Object` was
explicitly specified).

When the analyzer uses the common front end, it follows the spec, and
TypeParameterElement.bound returns a DartType of `Object` even if no
bound was specified.  To avoid this causing spurious test failures, we
need resynthesis tests to treat a bound of `Object` as equivalent to
an unspecified bound.

Change-Id: I3d09c2a1ee0be4273ef31ad4fe0ca4d92d7292cf
Reviewed-on: https://dart-review.googlesource.com/62723
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>