Revert separation of getOutline() and getResolution() in FrontEndCompiler.
It causes a problem when we first compile libraries A and B to outlines,
when A exports something from B. When we later compile B again to get
resolution, we exclude B from the list of dill libraries. This causes
DillLoader to fail in finalizeExports() because A.additionalExports
cannot symbols from B.

When we get back to compiling directily into resolved AST, we would
need more guidance from CFE team to help to resolve this problem.

R=brianwilkerson@google.com, paulberry@google.com

Change-Id: Ifb96a45999f6221e2e9af285d9d07b473d622e41
Reviewed-on: https://dart-review.googlesource.com/62666
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>