Fix a bug in top-level type inference dependency tracking
When there was a top-level type inference dependency from a declared
field to an overridden getter or setter, the dependency was not
recorded in the field.  This caused initializing formals to sometimes
fail to infer the field type, because they did not see that it needed
to be inferred.

Fixes https://github.com/dart-lang/sdk/issues/32866

Change-Id: If75658d087c099787f74af5f5c6db2ee6837febe
Reviewed-on: https://dart-review.googlesource.com/62800
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>