Analyzer/FE integration for writes to read-only variables.
Previously, when the user attempted to write to a read-only variable,
the front end creates a SyntheticExpressionJudgment to represent the
write.  This CL changes it to a new class,
InvalidVariableWriteJudgment, whose `infer` method sends the
appropriate resolution information to the analyzer.  Also, it ensures
that inference is performed on the RHS of the assignment, so that it
gets resolved too.

Partially addresses #33694.

Change-Id: I56b12e5082434ca8569d15df1b01e5aaad8f0a7d
Reviewed-on: https://dart-review.googlesource.com/63160
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>